Moves data.js to client/utils/
